requirements.txt
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.31.0
pyyaml>=6.0
pyarrow>=13.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pytest>=7.4.0
python-dateutil>=2.8.2
